requirements.txt
alembic==1.17.2
APScheduler==3.10.4
Authlib==1.6.5
babel==2.17.0
bcrypt==5.0.0
bleach==6.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==43.0.3
Deprecated==1.3.1
dotenv==0.9.9
Flask==2.3.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.0
idna==3.11
iso-639==0.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
lxml==5.2.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.2
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
netifaces-plus==0.12.5
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
pypdf==5.0.1
python-dotenv==1.2.1
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
regex==2024.5.15
requests==2.31.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.45
text-unidecode==1.3
tornado==6.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
Unidecode==1.3.8
urllib3==1.26.20
Wand==0.6.13
webencodings==0.5.1
Werkzeug==3.1.4
wrapt==2.0.1
WTForms==3.2.1
